品牌管理和类型管理后台代码

+++ luckygou-basic/basic-util/src/main/java/cn/itsource/luckygou/util/PageList.java
@@ -8,22 +8,22 @@ import java.util.List;
  * @param <T>
  */
 public class PageList<T> {
-    private Integer total = 0;
+    private Long total = 0L;
     private List<T> rows = new ArrayList<>();
 
     public PageList() {
     }
 
-    public PageList(Integer total, List<T> rows) {
+    public PageList(Long total, List<T> rows) {
         this.total = total;
         this.rows = rows;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
